+++ packages/tab/src/index.js
@@ -1,3 +1,1 @@
-import Tabs from './Tabs';
-
-export {Tabs};
+export { default as Tabs } from './Tabs';
